pyproject.toml
[project]
name = "spam-detect-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alith>=0.5.1",
    "pymilvus[model]>=2.5.6",
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.0",
]
